refactor: rename actor.rs to session.rs
